requirements.txt
Flask==2.3.3
gunicorn
werkzeug
dropbox==11.36.2
python-dotenv==1.0.0
requests==2.31.0
